Added rest repo for all components

+++ src/main/java/com/computerservice/repository/user/UserEntityRepository.java
@@ -3,6 +3,8 @@ package com.computerservice.repository.user;
 import com.computerservice.entity.user.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserEntityRepository extends JpaRepository<UserEntity, Integer> {
+import java.math.BigInteger;
+
+public interface UserEntityRepository extends JpaRepository<UserEntity, BigInteger> {
     UserEntity findByLogin(String login);
 }
